Cleaned up warnings

+++ test/test_emu_rom.c
@@ -13,7 +13,7 @@ static void EMUROM_testFileReadOk(void)
   uint8_t testRomContents[] = TEST_ROM_CONTENTS;
 
   uint8_t romBuf[30];
-  uint32_t romSize;
+  uint16_t romSize;
 
   TEST_ASSERT_INT_EQ(EMUROM_OK, EMUROM_read("TestRom.nes", romBuf, &romSize));
   TEST_ASSERT_INT_EQ(25u, romSize);
@@ -23,7 +23,7 @@ static void EMUROM_testFileReadOk(void)
 static void EMUROM_testFileReadNok(void)
 {
   uint8_t romBuf[1];
-  uint32_t romSize;
+  uint16_t romSize;
 
   TEST_ASSERT_UINT_EQ(EMUROM_NOK, EMUROM_read("ThisShouldNotWork.nes", romBuf, &romSize));
 }
